Additional tasks: Async, Swagger, separated codebase
